+++ NEWS.md
@@ -6,6 +6,7 @@
 
 ## New features
 
+- `common_language()` for estimating common language effect sizes.
 - `rbs_to_common_language()` for converting rank-biserial correlation to Probability of superiority.
 
 ## Changes